emailservice/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
mysqlclient==1.3.10
python-dotenv==0.10.3
python-http-client==3.2.1
pytz==2019.2
sendgrid==6.1.0
six==1.12.0
sqlparse==0.3.0
